Fixes, DTO.Aluno, DTO.Aministrador

+++ BLL/AcessoDB/Usuario.cs
@@ -3,6 +3,7 @@ using LinqToDB;
 using System.Linq;
 using System.Threading.Tasks;
 using System;
+using DTO;
 
 namespace BLL
 {
@@ -38,8 +39,7 @@ namespace BLL
             using (var db = new DTO.Database())
             {
                 var q = from u in db.TB_USUARIOS
-                        where u.ID == ID ||
-                              u.MATRICULA == MATRICULA ||
+                        where u.MATRICULA == MATRICULA ||
                               u.SENHA == SENHA ||
                               u.ID_PERFIL == ID_PERFIL
                         select u;
@@ -65,28 +65,10 @@ namespace BLL
             }
             return usr;
         }
-
-        /// <summary>
-        /// Procura um <typeparamref name="Usuario"/> no banco de dados de acordo com o id fornecido
-        /// </summary>
-        /// <param name="id">Id do <typeparamref name="Usuario"/> a ser consultado.</param>
-        /// <returns>Um <typeparamref name="Task"/> com o <typeparamref name="Usuario"/> encontrado.</returns>
-        public async Task<DTO.Usuario> ConsultarPorId(long id)
-        {
-            DTO.Usuario usr = null;
-            using (var db = new DTO.Database())
-            {
-                var q = from u in db.TB_USUARIOS
-                        where u.ID == id
-                        select u;
-                usr = await q.FirstOrDefaultAsync();
-            }
-            return usr;
-        }
         #endregion
 
         #region Métodos de MANIPULAÇÃO de dados
-        public void Inserir(DTO.Usuario u, DTO.Perfil p)
+        public void Inserir(DTO.Usuario u, ref DTO.Perfil p)
         {
 
             ///* RETIRANDO OS ESPAÇOS */
@@ -99,16 +81,24 @@ namespace BLL
             ///* HASH DA SENHA */
             u.SENHA = DBElementHandling.Hash(u.SENHA);
 
-            ///* TENTAR INSERIR O PERFIL */
-            Perfil pbll = new Perfil();
-            try { pbll.Inserir(p); } catch (Exception ex) { throw ex; }
+            // TENTAR INSERIR O PERFIL
+            if (u.ISADM)
+            {
+                Administrador abll = new Administrador();
+                try { abll.Inserir(p as DTO.Administrador); } catch (Exception ex) { throw ex; }
+            }
+            else
+            {
+                Aluno abll = new Aluno();
+                try { abll.Inserir(p as DTO.Aluno); } catch (Exception ex) { throw ex; }
+            }
 
             ///* CONECTANDO AO DB */
             DAL.Database db = new DAL.Database();
             db.Insert(u);
         }
 
-        public void Alterar(DTO.Usuario u, DTO.Perfil p, bool hashSenha = true)
+        public void Alterar(DTO.Usuario u, ref DTO.Perfil p, bool hashSenha = true)
         {
             // RETIRANDO OS ESPAÇOS
             DBElementHandling.RemoverEspacos(u);
@@ -121,9 +111,16 @@ namespace BLL
                 u.SENHA = DBElementHandling.Hash(u.SENHA);
 
             // TENTAR ALTERAR O PERFIL
-            Perfil pbll = new Perfil();
-
-            try { pbll.Alterar(p); } catch (Exception ex) { throw ex; }
+            if (u.ISADM)
+            {
+                Administrador abll = new Administrador();
+                try { abll.Alterar(p as DTO.Administrador); } catch (Exception ex) { throw ex; }
+            }
+            else
+            {
+                Aluno abll = new Aluno();
+                try { abll.Alterar(p as DTO.Aluno); } catch (Exception ex) { throw ex; }
+            }
 
             // CONECTANDO AO DB
             DAL.Database db = new DAL.Database();
@@ -138,13 +135,21 @@ namespace BLL
 
             u = await ConsultarPorMatricula(u.MATRICULA);
 
-            if (u == null) { throw new Exception("Usuário não encontrado"); };
+            if (u == null) { throw new Exception("Usuário não encontrado"); }
 
-            Perfil pbll = new Perfil();
-
-            DTO.Perfil p = await pbll.ConsultarPorId(u.ID_PERFIL);
-
-            try { pbll.Remover(p); } catch (Exception ex) { throw ex; }
+            // TENTAR ALTERAR O PERFIL
+            if (u.ISADM)
+            {
+                Administrador abll = new Administrador();
+                DTO.Administrador p = await abll.ConsultarPorID(u.ID_PERFIL);
+                try { abll.Remover(p); } catch (Exception ex) { throw ex; }
+            }
+            else
+            {
+                Aluno abll = new Aluno();
+                DTO.Aluno p = await abll.ConsultarPorID(u.ID_PERFIL);
+                try { abll.Remover(p); } catch (Exception ex) { throw ex; }
+            }
 
             DAL.Database db = new DAL.Database();
             db.Delete(u);
@@ -166,6 +171,22 @@ namespace BLL
         {
             if (u.MATRICULA == "") { throw new Exception("Matrícula inválida"); }
             if (u.SENHA == "") { throw new Exception("Senha inválida"); }
+        }
+
+        public async Task<Perfil> GetPerfil(DTO.Usuario u)
+        {
+            Perfil p = null;
+            if (u.ISADM)
+            {
+                Administrador abll = new Administrador();
+                p = await abll.ConsultarPorID(u.ID_PERFIL);
+            }
+            else
+            {
+                Aluno abll = new Aluno();
+                p = await abll.ConsultarPorID(u.ID_PERFIL);
+            }
+            return p;
         }
         #endregion
     }
